graph-rag-main/apps/content_scraper/requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.1
pydantic-settings==2.1.0

# Web scraping
playwright==1.40.0
beautifulsoup4==4.12.3
html2text==2020.1.16
aiohttp==3.9.3
requests==2.31.0
fake-useragent==1.5.1

# Database
gino==1.0.1
SQLAlchemy==1.3.24
psycopg2-binary==2.9.9

# Task queue
celery[redis]==5.3.6
redis==5.0.1
flower==2.0.1

# LangChain dependencies
langchain==0.2.10
langchain-openai==0.1.17
langchain-community==0.2.9
openai==1.12.0

# Utilities
tqdm==4.66.1
pytz==2024.1
environs==11.0.0
python-dotenv==1.0.1

# Data processing
scikit-learn==1.4.2
numpy==1.26.4

# Development tools
devtools==0.12.2

# Optional dependencies (commented out)
# uule_grabber==0.1.9
# apify-client==1.6.4  # Removed as Reddit/Apify support was removed
# aioredis==2.0.1  # Deprecated, use redis[hiredis] instead
# watchtower==3.2.0  # AWS CloudWatch logging